Move output of station.arrow
- also output day files and symlink to month/year
