+++ lib/bin/rollup.ts
@@ -2,26 +2,17 @@ import {mapAllCapped} from './mapallcapped';
 
 import {cloneDeep as _clonedeep, isEqual as _isequal, map as _map, reduce as _reduce, sortBy as _sortBy, filter as _filter, uniq as _uniq} from 'lodash';
 
-import {writeFileSync, unlinkSync, symlinkSync} from 'fs';
-import {createWriteStream} from 'fs';
-import {PassThrough} from 'stream';
-import {Utf8, Uint32, Float32, makeBuilder, Table, RecordBatchWriter} from 'apache-arrow/Arrow.node';
-
 import {rollupDatabase, purgeDatabase, rollupStartup, RollupDatabaseArgs} from '../worker/rollupworker';
 
 import {allStationsDetails, updateStationStatus} from './stationstatus';
 
 import {Accumulators, getCurrentAccumulators, AccumulatorTypeString, describeAccumulators} from './accumulators';
 
-import {gzipSync, createGzip} from 'zlib';
-
 import {Epoch, StationId} from './types';
 
 import {
     MAX_SIMULTANEOUS_ROLLUPS, //
-    STATION_EXPIRY_TIME_SECS,
-    OUTPUT_PATH,
-    UNCOMPRESSED_ARROW_FILES
+    STATION_EXPIRY_TIME_SECS
 } from '../common/config';
 
 export interface RollupStats {
@@ -210,7 +201,7 @@ export async function rollupAll(accumulators: Accumulators, nextAccumulators?: A
         );
 
         // And the global json
-        produceStationFile();
+        produceStationFile(accumulators);
 
         // Wait for all to be done
         await promise;
@@ -268,92 +259,6 @@ export async function rollupStartupAll() {
 
     startupStats.elapsed = Date.now() - now;
     console.log(`done initial rollup ${JSON.stringify(startupStats)}`);
-}
-
-//
-// Dump the meta data for all the stations, we take from our in memory copy
-// it will have been primed on start from the db and then we update and flush
-// back to the disk
-function produceStationFile() {
-    const stationDetailsArray = allStationsDetails();
-    // Form a list of hashes
-    let statusOutput = stationDetailsArray.filter((v) => {
-        return v.valid && v.lastPacket;
-    });
-
-    // Write this to the stations.json file
-    try {
-        const output = JSON.stringify(statusOutput);
-        writeFileSync(OUTPUT_PATH + 'stations.json', output);
-        writeFileSync(OUTPUT_PATH + 'stations.json.gz', gzipSync(output));
-    } catch (err) {
-        console.log('stations.json write error', err);
-    }
-
-    // Create an arrow version of the stations list - it will be smaller and quicker to
-    // load
-    try {
-        const id = makeBuilder({type: new Uint32()}),
-            name = makeBuilder({type: new Utf8()}),
-            lat = makeBuilder({type: new Float32()}),
-            lng = makeBuilder({type: new Float32()});
-
-        // Output an id sorted list of stations
-        for (const station of statusOutput.sort((a, b) => a.id - b.id)) {
-            id.append(station.id);
-            name.append(station.station);
-            lat.append(station.lat);
-            lng.append(station.lng);
-        }
-
-        // Convert into output file
-        const arrow = {
-            id: id.finish().toVector(),
-            name: name.finish().toVector(),
-            lat: lat.finish().toVector(),
-            lng: lng.finish().toVector()
-        };
-        const tableUpdates = new Table(arrow);
-
-        // And write them out
-        if (UNCOMPRESSED_ARROW_FILES) {
-            const pt = new PassThrough({objectMode: true});
-            pt.pipe(RecordBatchWriter.throughNode()) //
-                .pipe(createWriteStream(OUTPUT_PATH + 'stations.arrow'));
-            pt.write(tableUpdates);
-            pt.end();
-        }
-        {
-            const pt = new PassThrough({objectMode: true, emitClose: true});
-            pt.pipe(RecordBatchWriter.throughNode())
-                .pipe(createGzip())
-                .pipe(createWriteStream(OUTPUT_PATH + 'stations.arrow.gz'));
-            pt.write(tableUpdates);
-            pt.end();
-        }
-    } catch (error) {
-        console.log('stations.arrow write error', error);
-    }
-
-    // Write this to the stations.json file
-    try {
-        const output = JSON.stringify(stationDetailsArray);
-        writeFileSync(OUTPUT_PATH + 'stations-complete.json', output);
-        writeFileSync(OUTPUT_PATH + 'stations-complete.json.gz', gzipSync(output));
-    } catch (err) {
-        console.log('stations-complete.json write error', err);
-    }
-}
-
-export function symlink(src: string, dest: string) {
-    try {
-        unlinkSync(dest);
-    } catch (e) {}
-    try {
-        symlinkSync(src, dest, 'file');
-    } catch (e) {
-        console.log(`error symlinking ${src}.arrow to ${dest}: ${e}`);
-    }
 }
 
 function superThrow(t: string): never {
